Refactor toolbar into custom hook
